cxx compiler option + style
