+++ build.sh
@@ -1,8 +1,6 @@
 #!/bin/sh
-
-source ./config.sh
+. ./config.sh
 mkdir -p build
 cd build
-cmake -DUSER_CXX_FLAGS="${CXX_FLAGS}" ..
+cmake -DCMAKE_CXX_COMPILER="${CXX_COMPILER}" -DUSER_CXX_FLAGS="${CXX_FLAGS}" ..
 make
-cd ..
