+++ config_template.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
+#C++ compiler
+CXX_COMPILER="g++"
+
 # C++ flags
 CXX_FLAGS="-O3 -flto -Wall -Wno-unused-function"
 
-# currently unused:
-# C++ compiler
-# CXX="g++"
