+++ examples/ex2/run.sh
@@ -1,5 +1,3 @@
-#!/bin/bash
-ORIGDIR=$(pwd)
+#!/bin/sh
 cd ../../build/examples
 mpirun -np $1 ./ex2.exe
-cd "${ORIGDIR}"
