+++ test/CMakeLists.txt
@@ -1,3 +1,4 @@
+include_directories(common/)
 link_libraries(mci)
 add_executable(check main.cpp)
 add_executable(ut1.exe ut1/main.cpp)
