+++ test/ut3/main.cpp
@@ -5,37 +5,9 @@
 #include <math.h>
 #include <assert.h>
 
+#include "TestMCIFunctions.hpp"
+
 using namespace std;
-
-
-
-class ThreeDimGaussianPDF: public MCISamplingFunctionInterface{
-public:
-    ThreeDimGaussianPDF(): MCISamplingFunctionInterface(3, 1){}
-    ~ThreeDimGaussianPDF(){}
-
-    void samplingFunction(const double *in, double * protovalues){
-        protovalues[0] = (in[0]*in[0]) + (in[1]*in[1]) + (in[2]*in[2]);
-    }
-
-
-    double getAcceptance(const double * protoold, const double * protonew){
-        return exp(-protonew[0]+protoold[0]);
-    }
-
-};
-
-
-class XSquared: public MCIObservableFunctionInterface{
-public:
-    XSquared(): MCIObservableFunctionInterface(3, 1){}
-    ~XSquared(){}
-
-    void observableFunction(const double * in, double * out){
-        out[0] = in[0] * in[0];
-    }
-};
-
 
 
 int main(){
@@ -49,6 +21,8 @@ int main(){
     mci->setSeed(5649871);
     mci->addSamplingFunction(pdf);
     mci->addObservable(obs);
+    mci->setNBlocks(0); // use auto-blocking
+
     // the integral should provide 0.5 as answer!
 
     double * x = new double[3];
